add a general purpose storage area on the design system
This will allow us to add data to the design system itself instead of
using module-scope Map's where the key is pointing to the DesignSystem.

This has the benefit that all data is tied to a DesignSystem. If you
don't need a design system anymore, then garbage collection will kick
in and remove everything. If we rely on module scope maps, then we have
to do the cleanup (or use WeakMap, but that requires more testing to
know if that _actually_ works when multiple WeakMap instances reference
a DesignSystem).

Since it's a general purpose object, we have to ensure that you can't
accidentally override data... because of this, the storage is typed such
that the keys must be Symbols.