start preparing the design system
We need to enhance the design system to setup the storage on the design
system itself. We can use the `prepareDesignSystemStorage` to "promote"
the current `BaseDesignSystem` to the new `DesignSystem` where the
storage is properly setup.

This way, TypeScript will only allow you to access the necessary storage
parts if you are operating on a DesignSystem that has the storage setup
already.

Now it's up to us to ensure the storage is setup properly in the
`prepareDesignSystemStorage` function.