use `designSystem` as top-level DefaultMap key
Instead of nesting the `designSystem` as part of the nested signature
options object, we move the design system up. This commit is just here
to make the easy change later.

We want to store all caches on the design system itself, so by hoisting
it up, we can make a nicer change in the near future.