combine based on similar properties without values
We were combining utilities that were setting the same property _and_
the same value before testing.

This is correct for most of the utilities, but it will break down the
moment you want to combine values like this:

```
text-sm/6 leading-7
```

The `text-sm/6` would set a `line-height: 24px` but the `leading-7`
would set a `line-height: 28px`. Yet, if we combine both utilities the
`leading-7` will win, and the winning combination will become `text-sm/7`.

While this will lead to more combinations to try, the list will still be
limited to a handful.