chore: refresh lockfile
pnpm has got itself in a state it can't seem to get out of.

basically, it has 3 versions of esbuild install (0.23.x, 0.27.x, 0.25.x)
thanks to the fact we have `tsx` installed.

we don't depend on `tsx` anywhere, nor do any dependencies in the entire
tree. pnpm is holding onto it for no reason.

it is an _optional_ peer of vite and postcss-load-config (via tsup), one
which we don't depend on anywhere and so it shouldn't be installed.

I tried many different pnpm commands, configs, etc. Nothing seems to get
pnpm to behave here.

**Removing the lockfile entirely does solve the problem.**

It does mean some dependencies were upgraded, but mostly just esbuild
patches and what not.